auth-service/requirements.txt
fastapi~=0.112.0
PyJWT~=2.9.0
cryptography~=43.0.0
uvicorn~=0.30.5
franklin-fastapi-extension~=1.1.6

pydantic~=2.8.2
pydantic[email]~=2.8.2
starlette~=0.37.2
sqlalchemy~=2.0.0
pymysql~=1.0.2

PyJWT~=2.9.0
passlib~=1.7.4
cryptography~=43.0.0

bcrypt~=4.2.0